Generalize bundling attribute
The scope of this attribute goes beyond the Java ecosystem. So it is
defined at the same level as the usage attribute.

Issue #8380
